Add option to perform HCG with online fragment library
